ENH: Read and store "START" and "END" events from EDF file (#22)
* ENH: Read and store "START" and "END" events from EDF file

These events indicate when the Eyetracking started and stopped recording, respectively. There can be multiple of these events, meaning that the eyetracker started and paused (e.g. during a calibration sequence or between trials).

The pyeparse authors appeared to know about "START" and "END events in EDF files, but they just used python's `pass` to skip handling these events.

I adjusted our event handlers to give both start and end events an entry in `EDF["discrete"].

It's important to store this info because for some recordings (read: users) there is a time gap between an end event and the subsequent start event. In other words, there is a pause/gap in the recording that the user might want to know about.

* TST: add a unit test for start and end events